Fix file extension check on input file
Previously, the code checked for a '.' character in the name of the
input file. However, it was not checked if something followed upon
that dot, which would result in false positive passes on files that
do not actually have an existing file extension (e.g. "Example.",
"Video.2.1.").

Fix this by checking if there is a character followed by the last
dot found